move benchmark example app code to seperate file
